requirements.txt
absl-py==0.7.1
alabaster==0.7.12
allennlp==0.8.4
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.184
Babel==2.7.0
blis==0.2.4
boto3==1.9.151
botocore==1.12.174
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
conllu==0.11
cycler==0.10.0
cymem==2.0.2
docutils==0.14
editdistance==0.5.3
en-core-web-sm==2.1.0
flaky==3.5.3
Flask==1.0.3
Flask-Cors==3.0.8
ftfy==5.5.1
gast==0.2.2
gevent==1.4.0
greenlet==0.4.15
grpcio==1.20.1
h5py==2.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonnet==0.13.0
jsonpickle==1.2
jsonschema==2.6.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mock==3.0.5
more-itertools==7.0.0
munch==2.3.2
murmurhash==1.0.2
neuralcoref==4.0
nltk==3.4.3
numpy==1.16.3
numpydoc==0.9.1
overrides==1.9
packaging==19.0
parsimonious==0.8.1
plac==0.9.6
pluggy==0.12.0
preshed==2.0.1
protobuf==3.7.1
py==1.8.0
pyasn1==0.4.5
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.6.3
python-dateutil==2.8.0
pytorch-ignite==0.2.0
pytz==2019.1
PyYAML==5.1
regex==2019.4.14
requests==2.22.0
responses==0.10.6
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
snowballstemmer==1.2.1
spacy==2.1.3
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.0.5
tensorboard==1.13.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
thinc==7.0.4
torch==1.1.0
tqdm==4.32.1
Unidecode==1.1.1
urllib3==1.25.2
wasabi==0.2.2
wcwidth==0.1.7
Werkzeug==0.15.4
word2number==1.1
zipp==0.5.1
